Fix pars output from inertia sensor from double to int.

+++ src/autonomus.cpp
@@ -43,7 +43,7 @@ void Autonomus::writeAutonomusStream() {
   counter++;
   if (counter > 100) {
     std::ofstream ofs("/usd/autonomusStream.hpp", std::ios::app);
-    ofs << inertial.get_heading() << ", \t"
+    ofs << (int)inertial.get_heading() << ", \t"
         << leftFrontDiveTrainMotor.get_voltage() << ",\t"
         << rightFrontDiveTrainMotor.get_voltage() << ",\t"
         << launcherAMotor.get_voltage() << ",\t"
